requirements.txt
flask>=2.0
gunicorn>=20.0
pydantic>=1.10
requests>=2.0
Flask==2.3.3
folium==0.14.0
openai>=1.12.0
requests==2.31.0
python-dotenv==1.0.0

# CORS support for serving API to separate frontend origins
Flask-Cors>=3.0.10

# WebSocket support for real-time progress updates
flask-socketio>=5.3.0
python-socketio>=5.10.0
eventlet>=0.33.0

# Redis cache support for performance optimization
redis>=5.0.0
hiredis>=2.2.0

# Database support
# We install components separately to avoid the `storage3` dependency which requires a C++ compiler.
gotrue>=2.0.0
postgrest>=0.12.0
realtime>=0.10.0
sqlalchemy
